--- ReadMe.md
@@ -1,39 +1,0 @@
-﻿MIC Certificate Generator
-
-
-- Summary:
-
-- The idea of this project is to create a unified web app that allows us to create an event with brandings and authorized signatory for the specific event and generate bulk certificates for all participants and/or winners.
-
-Repo Link: <https://github.com/micvitc/Certificate-Generator/>
-
-- The relevant resources will be sent in the groups.
-
-- Frameworks:
-
-
-
-- Web Dev:
-  - Poetry/Venv
-  - Flask
-  - Flask-RESTful
-  - Flask-pyMongo
-  - OpenCV
-  - Pandas
-
-- WorkFlow and Requirements:
-
-- Web Dev:
-  - Initialize the environment setups and Initial Flask App to Display a home screen and a navbar button for a login screen
-  - Start Constructing models or schemas for database and submit ER Diagrams and Activity diagrams.
-  - Design a OpenCV engine that allows you to create and alter certificate templates according to the design and can be called with APIs for specific attribute changes.
-  - Dashboard to create new event, add templates, authorized signatories, branding, winners, participants, etc.
-  - Both the teams will make the same web app separately. Although, you are allowed to consult the other team to gain ideas from each other.
-
-- Management:
-  - Document the up-to-date changes depending on the roll-out of tasks.
-  - Each management will be assigned technical tasks as well which they should be able to collaborate with the respective departments.
-
-
-
-
