--- data.csv
@@ -1,10 +1,0 @@
-1,John Doe,First
-1,Jane Smith,Second
-2,Michael Johnson,Third
-2,Sarah Williams,Participation
-3,David Brown,First
-3,Emily Davis,Second
-4,Andrew Taylor,Third
-4,Olivia Anderson,Participation
-5,James Wilson,First
-5,Sophia Martinez,Second
