+++ requirements.txt
@@ -1,0 +1,9 @@
+Flask==2.2.5
+flask-marshmallow==0.15.0
+Flask-SQLAlchemy==3.0.5
+Jinja2==3.1.2
+MarkupSafe==2.1.3
+marshmallow==3.19.0
+opencv-python==4.7.0.72
+SQLAlchemy==2.0.17
+gunicorn==20.1.0
