--- static/update2.css
@@ -1,143 +1,0 @@
-body {
-    background-image: linear-gradient(135deg, #ee00ff 10%, #1904E5 100%);
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-attachment: fixed;
-    font-family: "Open Sans", sans-serif;
-    color: #333333;
-  }
-  
-  .box-form {
-    margin: 0 auto;
-    width: 80%;
-    background: #ffffff;
-    border-radius: 10px;
-    overflow: hidden;
-    display: flex;
-    flex: 1 1 100%;
-    align-items: stretch;
-    justify-content: space-between;
-    box-shadow: 0 0 20px 6px #090b6f85;
-  }
-  @media (max-width: 980px) {
-    .box-form {
-      flex-flow: wrap;
-      text-align: center;
-      align-content: center;
-      align-items: center;
-    }
-  }
-  .box-form div {
-    height: auto;
-  }
-  .box-form .left {
-    color: #FFFFFF;
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-image: url("bg.jpg");
-    overflow: hidden;
-  }
-  .box-form .left .overlay {
-    padding: 30px;
-    width: 100%;
-    height: 100%;
-    background: #9159f9ad;
-    overflow: hidden;
-    box-sizing: border-box;
-  }
-  .box-form .left .overlay h1 {
-    font-size: 6vmax;
-    line-height: 1;
-    font-weight: 600;
-    margin-top: 150px;
-    margin-bottom: 20px;
-    color: white;
-    font-style:bold ;
-    text-align: center;
-  }
-  .box-form .left .overlay span p {
-    margin-top: 30px;
-    font-weight: 900;
-    text-align: center;
-  }
-  .box-form .left .overlay span a {
-    background: #3b5998;
-    color: #FFFFFF;
-    margin-top: 10px;
-    padding: 14px 50px;
-    border-radius: 100px;
-    display: inline-block;
-    box-shadow: 0 3px 6px 1px #042d4657;
-  }
-  .box-form .left .overlay span a:last-child {
-    background: #1dcaff;
-    margin-left: 30px;
-  }
-  .box-form .right {
-    padding: 40px;
-    overflow: hidden;
-  }
-  @media (max-width: 980px) {
-    .box-form .right {
-      width: 100%;
-    }
-  }
-  .box-form .right h5 {
-    font-size: 6vmax;
-    line-height: 0;
-  }
-  .box-form .right p {
-    font-size: 14px;
-    color: #B0B3B9;
-  }
-  .box-form .right .inputs {
-    overflow: hidden;
-  }
-  .box-form .right input {
-    width: 100%;
-    padding: 10px;
-    margin-top: 25px;
-    font-size: 16px;
-    border: none;
-    outline: none;
-    border-bottom: 2px solid #B0B3B9;
-  }
-
-  .box-form .right button {
-    float: right;
-    color: #fff;
-    font-size: 16px;
-    padding: 12px 35px;
-    border-radius: 50px;
-    display: inline-block;
-    border: 0;
-    outline: 0;
-    box-shadow: 0px 4px 20px 0px #49c628a6;
-    background-image: linear-gradient(135deg, #70F570 10%, #49C628 100%);
-  }
-  
-  label {
-    display: block;
-    position: relative;
-    margin-left: 30px;
-  }
-  
-  label::before {
-    content: ' \f00c';
-    position: absolute;
-    font-family: FontAwesome;
-    background: transparent;
-    border: 3px solid #70F570;
-    border-radius: 4px;
-    color: transparent;
-    left: -30px;
-    transition: all 0.2s linear;
-  }
-  
-  label:hover::before {
-    font-family: FontAwesome;
-    content: ' \f00c';
-    color: #fff;
-    cursor: pointer;
-    background: #70F570;
-  }
