--- templates/Website Final1.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Home Page</title>
-  <link href="https://fonts.googleapis.com/css?family=Anton" rel="stylesheet">
-  <link rel="stylesheet" href="{{ url_for('static', filename='update2.css') }}" />
-  <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
-</head>
-<body>
-<div class="box-form">
-    <div class="left">
-        <div class="overlay">
-            <h1>CERTIFICATE<br> GENERATOR</h1>
-            <p align="center">"Empowering Recognition: Create and Customize Certificates Online"</p>
-        </div>
-    </div>
-    <div class="right">
-        <h5>Login</h5>
-        <p>Don't have an account? <a href="{{ url_for('adminsignup') }}">Create Your Account</a> it takes less than a minute</p>
-        <form method="POST" action="{{ url_for('adminlogin') }}">
-            <div class="inputs">
-                <input type="text" name="username" placeholder="Username">
-                <br>
-                <input type="password" name="password" placeholder="Password">
-            </div>
-            
-            <br>
-            <button type="submit">Login</button>
-        </form>
-
-        {% with messages = get_flashed_messages(with_categories=true) %}
-          {% if messages %}
-            {% for category, message in messages %}
-              {% if category == 'error' %}
-                <script>
-                  alert("{{ message }}");
-                </script>
-              {% endif %}
-            {% endfor %}
-          {% endif %}
-        {% endwith %}
-        
-        
-    </div>
-</div>
-</body>
-</html>
