--- templates/admin.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Admin Page - Certificate Generator</title>
-  <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.16/dist/tailwind.min.css" rel="stylesheet">
-  <style>
-    .navbar-gradient {
-      background-image: linear-gradient(to left, #f200ff, #8800ff);
-    }
-  </style>
-</head>
-<body class="bg-clip-padding-20">
-    <nav class="fixed top-0 z-50 w-full bg-gradient-to-r from-pink-500 to-purple-500 border-b border-gray-200 dark:bg-gray-800 dark:border-gray-700">
-      <!-- Navbar content -->
-    </nav>
-  
-<main class="flex-center items-center m-20">
-    <div class="flex-2">
-      <div class="mx-auto my-8 p-8 bg-gray-200 rounded-2xl">
-        <h2 class="text-2xl font-bold uppercase italic text-center">User Data</h2>
-        
-        <table class="w-full mt-6 border-collapse">
-          <thead>
-            <tr>
-              <th class="p-3 font-bold uppercase bg-purple-200 text-gray-600 border border-gray-300 hidden lg:table-cell">User</th>
-              <th class="p-3 font-bold uppercase bg-purple-200 text-gray-600 border border-gray-300 hidden lg:table-cell">Password</th>
-              <th class="p-3 font-bold uppercase bg-purple-200 text-gray-600 border border-gray-300 hidden lg:table-cell">Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            {% for user in users %}
-              <tr class="hover:bg-purple-100 border-b border-gray-200 py-10">
-                <td class="p-3 text-gray-600 border border-gray-300 hidden lg:table-cell">{{ user.username }}</td>
-                <td class="p-3 text-gray-600 border border-gray-300 hidden lg:table-cell">{{ user.password }}</td>
-                <td class="p-3 text-gray-600 border border-gray-300 hidden lg:table-cell">
-                  <form action="{{ url_for('delete_user') }}" method="POST">
-                    <input type="hidden" name="username" value="{{ user.username }}">
-                    <button type="submit" class="text-red-600">Delete</button>
-                  </form>
-                </td>
-              </tr>
-            {% endfor %}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  </main>
-</body>
-</html>
