--- templates/generator.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Generator - Certificate Generator</title>
-  <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.16/dist/tailwind.min.css" rel="stylesheet">
-  <style>
-    body {
-      background: linear-gradient(to bottom left,#fefcfe, #f6f4f6, #faf9fa,#f2b8ff);
-    }
-    .navbar-gradient {
-      background-image: linear-gradient(to left, #f200ff, #8800ff);
-    }
-    .sign-out-button {
-      background-color: #48BB78;
-    }
-  </style>
-    <script>
-      function confirmSignOut() {
-        if (confirm("Are you sure you want to sign out?")) {
-          window.location.href = "{{ url_for('home') }}"; 
-        }
-      }
-    </script>
-</head>
-<body class="bg-clip-padding-20 bg-fixed">
-  <nav class="fixed top-0 z-50 w-full bg-gradient-to-r from-pink-500 to-purple-500 border-b border-gray-200 dark:bg-gray-800 dark:border-gray-700">
-    <div class="px-3 py-3 lg:px-5 lg:pl-3">
-      <div class="flex items-center justify-between">
-        <div class="flex items-center justify-start">
-          <button data-drawer-target="logo-sidebar" data-drawer-toggle="logo-sidebar" aria-controls="logo-sidebar" type="button" class="inline-flex items-center p-2 text-sm text-gray-500 rounded-lg sm:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600">
-            <span class="sr-only">Open sidebar</span>
-            <svg class="w-6 h-6" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-              <path clip-rule="evenodd" fill-rule="evenodd" d="M2 4.75A.75.75 0 012.75 4h14.5a.75.75 0 010 1.5H2.75A.75.75 0 012 4.75zm0 10.5a.75.75 0 01.75-.75h7.5a.75.75 0 010 1.5h-7.5a.75.75 0 01-.75-.75zM2 10a.75.75 0 01.75-.75h14.5a.75.75 0 010 1.5H2.75A.75.75 0 012 10z"></path>
-            </svg>
-          </button>
-            <span class="self-center text-xl font-semibold uppercase sm:text-3xl whitespace-nowrap text-white">Certificate Generator</span>
-        </div>
-        <div class="flex items-center">
-          <div class="flex items-center ml-3">
-            <div>
-              <a href="https://your-sign-out-url.com" class="flex text-white rounded-full focus:ring-4 focus:ring-gray-300 dark:focus:ring-gray-600">
-                <span class="sr-only">Back</span>
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </nav>
-  
-  <div class="mt-40"></div>
-  
-  <div class="max-w-md mx-auto p-6 bg-gray-200 rounded-2xl">
-    <form action="{{ url_for('mainpage') }}" method="POST" enctype="multipart/form-data">
-      <div class="max-w-xl mx-auto p-6 bg-gray-200 rounded-2xl">
-        <div class="mb-4">
-          <h2 class="text-2xl font-semibold text-gray-700 mb-2">Add Participants from CSV</h2>
-          <hr class="mb-4">
-        </div>
-        <div class="mb-4">
-          <label for="csv_file" class="block mb-2 text-lg font-semibold text-gray-700">Choose CSV File:</label>
-          <input type="file" id="csv_file" name="csv_file" accept=".csv" class="w-full px-4 py-2 text-gray-700 border rounded-md focus:outline-none focus:ring-2 focus:ring-purple-500" required>
-        </div>
-        <button type="submit" class="w-full px-4 py-2 text-white bg-purple-500 rounded-md hover:bg-purple-600 focus:outline-none focus:ring-2 focus:ring-purple-500">Submit</button>
-      </div>
-      
-    </form>
-    <div class="mt-4 flex justify-center">
-      <button onclick="confirmSignOut()" class="px-4 py-2 text-white rounded-md sign-out-button">Back</button>
-    </div>
-  </div>
-
-</body>
-</html>
